requirements.txt
streamlit==1.21.0
openai==1.0.0
pyht==1.0.0
python-dotenv==1.0.0
requests==2.28.1